For all tests that require a DynamoViewModel (or symbols from DynamoCoreWpf) move to DynamoCoreWpfTests and also remove all PresentationCore, WindowsBase references from DynamoCoreTests
