+++ test/DynamoCoreTests/DynamoCoreTests.csproj
@@ -44,10 +44,6 @@
       <Private>False</Private>
     </Reference>
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="Microsoft.Practices.Prism, Version=4.1.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\..\extern\prism\Microsoft.Practices.Prism.dll</HintPath>
-    </Reference>
     <Reference Include="Moq, Version=4.2.1312.1622, Culture=neutral, PublicKeyToken=69f491c39445e920, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\..\extern\Moq\Moq.dll</HintPath>
@@ -56,24 +52,18 @@
       <SpecificVersion>False</SpecificVersion>
       <HintPath>$(NunitPath)\nunit.framework.dll</HintPath>
     </Reference>
-    <Reference Include="PresentationCore" />
-    <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xaml" />
     <Reference Include="System.XML" />
-    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AnnotationModelTest.cs" />
     <Compile Include="AstBuilderTest.cs" />
     <Compile Include="CallsiteTests.cs" />
-    <Compile Include="ConverterTests.cs" />
     <Compile Include="CrashProtectionTests.cs" />
     <Compile Include="DSEvaluationUnitTestBase.cs" />
     <Compile Include="DynamoModelTestBase.cs" />
     <Compile Include="MessageLogTests.cs" />
-    <Compile Include="IconsTests.cs" />
     <Compile Include="NodeConstructionTests.cs" />
     <Compile Include="PackageManager\MockMaker.cs" />
     <Compile Include="PackageManager\PackageDirectoryBuilderTests.cs" />
@@ -81,7 +71,6 @@
     <Compile Include="PackageManager\PackageTests.cs" />
     <Compile Include="PackageManager\RecordedFileSystem.cs" />
     <Compile Include="PackageValidationTest.cs" />
-    <Compile Include="SearchMemberGroupTests.cs" />
     <Compile Include="TestZeroTouchClass.cs" />
     <Compile Include="TypedParametersToStringTests.cs" />
     <Compile Include="CodeBlockNodeTests.cs" />
@@ -90,7 +79,6 @@
     <Compile Include="DSFunctionNodeTest.cs" />
     <Compile Include="DSLibraryTest.cs" />
     <Compile Include="DynamoDefects.cs" />
-    <Compile Include="ExecutionIntervalTests.cs" />
     <Compile Include="LibraryCustomizationTests.cs" />
     <Compile Include="LibraryTests.cs" />
     <Compile Include="MigrationManagerTests.cs" />
@@ -105,27 +93,20 @@
     <Compile Include="ScopedNodeTest.cs" />
     <Compile Include="PackageManager\PackageUtilitiesTests.cs" />
     <Compile Include="SchedulerTests.cs" />
-    <Compile Include="SearchSideEffects.cs" />
-    <Compile Include="SearchViewModelTests.cs" />
     <Compile Include="Setup.cs" />
     <Compile Include="UnicodeTest.cs" />
     <Compile Include="UnitTestBase.cs" />
     <Compile Include="UpdateManagerTests.cs" />
     <Compile Include="UserDataMigrationTests.cs" />
     <Compile Include="VersionUtilitiesTests.cs" />
-    <Compile Include="WatchNodeTests.cs" />
     <Compile Include="WorkspaceHeaderTests.cs" />
-    <Compile Include="WorkspaceOpeningTests.cs" />
-    <Compile Include="WorkspaceSaving.cs" />
     <Compile Include="CoreTests.cs" />
     <Compile Include="CoreDynTests.cs" />
     <Compile Include="CustomNodes.cs" />
     <Compile Include="FileReading.cs" />
     <Compile Include="Nodes\ListTests.cs" />
-    <Compile Include="DynamoViewModelUnitTest.cs" />
     <Compile Include="Nodes\LogicTests.cs" />
     <Compile Include="Nodes\StringTests.cs" />
-    <Compile Include="PackageManager\PackageDependencyTests.cs" />
     <Compile Include="PackageManager\PackageUploadBuilderTests.cs" />
     <Compile Include="SearchModelTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -137,11 +118,6 @@
     <Compile Include="XmlHelperTests.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\src\DynamoCoreWpf\DynamoCoreWpf.csproj">
-      <Project>{51bb6014-43f7-4f31-b8d3-e3c37ebedaf4}</Project>
-      <Name>DynamoCoreWpf</Name>
-      <Private>False</Private>
-    </ProjectReference>
     <ProjectReference Include="..\..\src\DynamoCore\DynamoCore.csproj">
       <Project>{7858FA8C-475F-4B8E-B468-1F8200778CF8}</Project>
       <Name>DynamoCore</Name>
