+++ test/DynamoCoreWpfTests/IconsTests.cs
