+++ test/DynamoCoreWpfTests/SearchMemberGroupTests.cs
