+++ test/DynamoCoreWpfTests/SearchSideEffects.cs
