+++ test/DynamoCoreWpfTests/SearchViewModelTests.cs
