+++ test/DynamoCoreWpfTests/WatchNodeTests.cs
